requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
pydantic>=2.10.0
pydantic-settings>=2.6.0
pyyaml>=6.0.1
python-dateutil>=2.8.2
aiosqlite>=0.20.0
aiomysql>=0.2.0
bcrypt>=4.0.0
python-jose[cryptography]>=3.3.0
pytest>=8.0.0
pytest-asyncio>=0.24.0
httpx>=0.28.0
